Moved AgendaType to the input model.
