+++ src/app/components/agenda-generator/agenda-model.ts
@@ -1,4 +1,5 @@
 import { Time } from '@angular/common';
+import { AgendaType } from 'src/app/services/agenda/agenda-type';
 import { DayPlanInput } from '../../services/day-plan-input';
 
 export class AgendaModel {
@@ -18,6 +19,7 @@ export class AgendaModel {
       tomeks: false,
       running: false,
       atOffice: atOfficeTimes[0],
+      agendaType: AgendaType.Default
     };
   }
 }
