+++ src/app/components/type-exclusive-agenda-generator/type-exclusive-agenda-generator.component.html
@@ -20,7 +20,7 @@
         type="radio"
         name="agenda-type"
         [value]="aType"
-        [(ngModel)]="agendaType"
+        [(ngModel)]="model.agendaInput.agendaType"
         id="{{ aType }}-agenda"
       />
       <label class="form-check-label" for="{{ aType }}-agenda"
