+++ src/app/components/type-exclusive-agenda-generator/type-exclusive-agenda-generator.component.ts
@@ -3,7 +3,7 @@ import { AgendaGenerationService } from '../../services/agenda-generation.servic
 import { AgendaSynchronizeService } from '../../services/agenda-synchronize.service';
 import { AtOfficeTimesProviderService } from '../../services/at-office-times-provider.service';
 import { AgendaModel } from '../agenda-generator/agenda-model';
-import { AgendaType } from './agenda-type';
+import { AgendaType } from '../../services/agenda/agenda-type';
 
 @Component({
   selector: 'app-type-exclusive-agenda-generator',
@@ -14,7 +14,6 @@ export class TypeExclusiveAgendaGeneratorComponent {
   Object = Object;
   model: AgendaModel;
   readonly agendaTypes: AgendaType[];
-  agendaType: AgendaType;
 
   constructor(
     private agendaService: AgendaGenerationService,
@@ -24,9 +23,13 @@ export class TypeExclusiveAgendaGeneratorComponent {
     const atOfficeTimes = atOfficeTimeService.generateAtOfficeSelection();
     this.model = new AgendaModel(atOfficeTimes);
 
+    this.agendaTypes = this.getAllAgendaTypes();
+    this.model.agendaInput.agendaType = this.agendaTypes[0];
+  }
+
+  private getAllAgendaTypes()  : AgendaType[] {
     const keys = Object.values(AgendaType).filter(k => typeof AgendaType[k as any] !== "number");
-    this.agendaTypes = keys.map((k) => AgendaType[k]);
-    this.agendaType = this.agendaTypes[0];
+    return keys.map((k) => AgendaType[k]);
   }
 
   onSubmit() {
