+++ src/app/services/day-plan-input.ts
@@ -1,8 +1,10 @@
 import { Time } from '@angular/common';
+import { AgendaType } from './agenda/agenda-type';
 
 export class DayPlanInput {
   label: string;
   morningPages: boolean;
+  agendaType: AgendaType;
   training: boolean;
   tomeks: boolean;
   running: boolean;
